release: 3.5.2
update: Newest Slick and Nifty version
fix: Doesn't crash while loading an old high score file
feat: Sound for the buttons in main menu
design: New main menu background image, new mouse cursor

+++ src/lib/mylib/color/Transparency.java
@@ -9,7 +9,7 @@ import org.newdawn.slick.Color;
 public class Transparency {
 
 	public static Color getTransparency(float alpha) {
-		return new Color(255, 255, 255, (float) alpha);
+		return new Color(255, 255, 255, alpha);
 	}
 
 	public static Color setTransparency(Color color, float alpha) {
